refactor(ms): readme, test, inmem handlers, eslint import rule (#5)
* test(feature-id): cover test with unique values

* refactor(ms): move in mem handlers to better place, maybe

* refactor(ms): add imports/export sort rule

* refactor(ms): better readme
