+++ package.json
@@ -55,6 +55,7 @@
     "eslint": "8.56.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-prettier": "5.1.3",
+    "eslint-plugin-simple-import-sort": "12.1.1",
     "husky": "9.1.4",
     "lint-staged": "15.2.7",
     "prettier": "3.3.3",
