+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from './handlers/http/ms-register-handler'
 export * from './mock-service'
-
-export { HttpResponse, passthrough, bypass, delay } from 'msw'
+export { bypass, delay, HttpResponse, passthrough } from 'msw'
