+++ src/ms-dashboard.ts
@@ -1,6 +1,9 @@
 import { StatusCodes } from 'http-status-codes'
 import { Readable } from 'stream'
 
+import { config } from '~/const'
+import { MSInMemHandler, MSInMemHandlers } from '~/core/in-mem-handlers'
+import { MSHttpResponse } from '~/handlers/http/response'
 import {
   IMockService,
   IMSDashboard,
@@ -13,12 +16,6 @@ import {
   MSTrackableRequestContent,
   MSTrackableRequestContentShort,
 } from '~/types'
-import {
-  MSInMemHandler,
-  MSInMemHandlers,
-} from '~/handlers/http/in-mem-handlers'
-import { MSHttpResponse } from '~/handlers/http/response'
-import { config } from '~/const'
 import { secToMinHuman } from '~/utils/time'
 
 /**
