+++ src/ms-request.ts
@@ -1,4 +1,5 @@
 import { DefaultBodyType, StrictRequest } from 'msw'
+
 import { bodyJSONParser } from '~/handlers/http/body-json-parser'
 import { IdResolver } from '~/id-resolver'
 import { MSTrackableRequestContent } from '~/types/ms-request.types'
