+++ src/types/index.ts
@@ -1,3 +1,3 @@
+export * from './ms.types'
 export * from './ms-request.types'
 export * from './ms-response.types'
-export * from './ms.types'
